move WMFAuthenticationManager to WMF
