+++ WMF Framework/WMFCaptchaViewController.swift
@@ -30,29 +30,6 @@ extension UIStackView {
     func captchaHideSubtitle() -> Bool
 }
 
-public class WMFCaptcha: NSObject {
-    @objc let captchaID: String
-    @objc let captchaURL: URL
-    @objc init(captchaID:String, captchaURL:URL) {
-        self.captchaID = captchaID
-        self.captchaURL = captchaURL
-    }
-    static func captcha(from requests: [[String : AnyObject]]) -> WMFCaptcha? {
-        guard
-            let captchaAuthenticationRequest = requests.first(where:{$0["id"]! as! String == "CaptchaAuthenticationRequest"}),
-            let fields = captchaAuthenticationRequest["fields"] as? [String : AnyObject],
-            let captchaId = fields["captchaId"] as? [String : AnyObject],
-            let captchaInfo = fields["captchaInfo"] as? [String : AnyObject],
-            let captchaIdValue = captchaId["value"] as? String,
-            let captchaInfoValue = captchaInfo["value"] as? String,
-            let captchaURL = URL(string: captchaInfoValue)
-            else {
-                return nil
-        }
-        return WMFCaptcha(captchaID: captchaIdValue, captchaURL: captchaURL)
-    }
-}
-
 class WMFCaptchaViewController: UIViewController, UITextFieldDelegate, Themeable {
 
     @IBOutlet fileprivate var captchaImageView: UIImageView!
